fix(store): exact 64-bit graph counts; stop truncating COUNT(*) at INT_MAX
SQLite computes COUNT(*) as a signed 64-bit integer, but finish_cached_count
read it with sqlite3_column_int and every public count API returned int. Each
node/edge count therefore wrapped at the signed 32-bit boundary, and a wrapped
value is indistinguishable from a genuine small one to every consumer:
artifact export/import equality, fleet reporting, MCP statistics, plausibility
gates, and kernel telemetry could all accept or publish a count that is not the
physical row cardinality.

The contract is now int64_t end to end:
  finish_cached_count            -> int64_t via sqlite3_column_int64
  cbm_store_count_nodes/_scoped  -> int64_t
  cbm_store_count_edges/_scoped  -> int64_t
  cbm_store_count_edges_by_type  -> int64_t
  cbm_store_count_vectors        -> int64_t
and through bindgen to Rust as i64. CBM_STORE_ERR (-1) stays the sole failure
sentinel, which remains unambiguous because a physical row count is never
negative -- no clamp, no wrap, no silent legacy fallback.

artifact.c is the case the issue named: write_metadata took int, so the source
query could wrap before the existing metadata range check ever ran. Both the
locals and the parameter are now int64_t, so the count reaching artifact
metadata is the measured one.

-Werror -Wformat caught the two remaining narrowings itself (http_server.c
health payload printed %d); both now print the exact value.

Verified on aarch64-apple-darwin, published counts against an independent
64-bit SQL read of the same store:
  cssrepro  published nodes=9  edges=10   sqlite nodes=9  edges=10
  scssns    published nodes=20 edges=23   sqlite nodes=20 edges=23
  absent project -> exactly 0
Type chain, greppable as a standing check:
  sqlite3_column_int on the count path : 0
  sqlite3_column_int64                 : 1
  count fns returning int64_t          : 6/6
  bindgen surface returning i64        : 6/6

The > INT_MAX case is unreachable by construction on this hardware: crossing it
needs 2,147,483,648 physical rows, roughly 20 GB of SQLite even for a minimal
table. It is closed by the type contract above rather than by materialisation,
and that is stated rather than implied.

Refs #952

+++ cbm/src/mcp/mcp.c
@@ -3185,8 +3185,8 @@ static db_project_inspect_status_t build_project_json_entry(cbm_mcp_server_t *sr
         return inspect;
     }
 
-    int nodes = cbm_store_count_nodes(pstore, project_name);
-    int edges = cbm_store_count_edges(pstore, project_name);
+    int64_t nodes = cbm_store_count_nodes(pstore, project_name);
+    int64_t edges = cbm_store_count_edges(pstore, project_name);
     if (nodes < 0 || edges < 0) {
         record_store_query_failure(srv, "", full_path, pstore, CBM_STORE_VERIFY_IO_FAILED,
                                    "source.query_project_details", cbm_store_error(pstore));
@@ -4808,8 +4808,8 @@ static char *handle_index_status(cbm_mcp_server_t *srv, const char *args) {
     yyjson_mut_doc_set_root(doc, root);
 
     if (project) {
-        int nodes = cbm_store_count_nodes(store, project);
-        int edges = cbm_store_count_edges(store, project);
+        int64_t nodes = cbm_store_count_nodes(store, project);
+        int64_t edges = cbm_store_count_edges(store, project);
         yyjson_mut_obj_add_str(doc, root, "project", project);
         yyjson_mut_obj_add_int(doc, root, "nodes", nodes);
         yyjson_mut_obj_add_int(doc, root, "edges", edges);
@@ -5080,8 +5080,8 @@ static char *handle_get_architecture(cbm_mcp_server_t *srv, const char *args) {
                                aspects_strs_count > 0 ? aspects_strs : NULL, aspects_strs_count,
                                &arch);
 
-    int node_count = cbm_store_count_nodes_scoped(store, project, scope_path);
-    int edge_count = cbm_store_count_edges_scoped(store, project, scope_path);
+    int64_t node_count = cbm_store_count_nodes_scoped(store, project, scope_path);
+    int64_t edge_count = cbm_store_count_edges_scoped(store, project, scope_path);
     char norm_path[CBM_SZ_512];
     bool path_scoped = cbm_store_normalize_arch_path(scope_path, norm_path, sizeof(norm_path));
 
@@ -5094,8 +5094,8 @@ static char *handle_get_architecture(cbm_mcp_server_t *srv, const char *args) {
     }
     if (path_scoped) {
         yyjson_mut_obj_add_str(doc, root, "path", norm_path);
-        int root_nodes = cbm_store_count_nodes(store, project);
-        int root_edges = cbm_store_count_edges(store, project);
+        int64_t root_nodes = cbm_store_count_nodes(store, project);
+        int64_t root_edges = cbm_store_count_edges(store, project);
         yyjson_mut_obj_add_int(doc, root, "root_total_nodes", root_nodes);
         yyjson_mut_obj_add_int(doc, root, "root_total_edges", root_edges);
         yyjson_mut_obj_add_int(doc, root, "scoped_total_nodes", node_count);
@@ -6371,8 +6371,8 @@ static char *build_index_success_response(cbm_mcp_server_t *srv, yyjson_mut_doc 
     if (!store) {
         return build_no_store_error(srv, project_name);
     }
-    int nodes = cbm_store_count_nodes(store, project_name);
-    int edges = cbm_store_count_edges(store, project_name);
+    int64_t nodes = cbm_store_count_nodes(store, project_name);
+    int64_t edges = cbm_store_count_edges(store, project_name);
     if (nodes < 0 || edges < 0) {
         record_store_query_failure(srv, project_name, cbm_store_db_path(store), store,
                                    CBM_STORE_VERIFY_IO_FAILED, "source.query_persisted_counts",
